services/data-upload/requirements.txt
fastapi
uvicorn
pandas
azure-storage-blob
